Moved AddonsItem component to addons.js file
